Split up interfaces into separate files
This solves a circular dependency issue
