+++ index.ts
@@ -1,3 +1,7 @@
 export { ModalConfig } from './src/modal-config';
-export { ModalManagerInterface, ModalManager } from './src/modal-manager';
+export { ModalManager } from './src/modal-manager';
+export { ModalManagerMode } from './src/modal-manager-mode';
+export { ModalManagerInterface } from './src/modal-manager-interface';
+export { ModalManagerHostBridge } from './src/modal-manager-host-bridge';
+export { ModalManagerHostBridgeInterface } from './src/modal-manager-host-bridge-interface';
 export { ModalTemplate } from './src/modal-template';
