+++ src/modal-manager-mode.ts
@@ -1,0 +1,10 @@
+/**
+ * Various modes the modal can be in
+ *
+ * @export
+ * @enum {number}
+ */
+export enum ModalManagerMode {
+  Open = 'open',
+  Closed = 'closed',
+}
